requirements.txt
wonderwords
google-cloud-core
bigquery
nicegui
db-dtypes